refactor: reduceLazy move to strict folder (#247)
