+++ src/Lazy/index.ts
@@ -24,7 +24,6 @@ import pipeLazy from "./pipeLazy";
 import pluck from "./pluck";
 import prepend from "./prepend";
 import range from "./range";
-import reduceLazy from "./reduceLazy";
 import reject from "./reject";
 import repeat from "./repeat";
 import reverse from "./reverse";
@@ -71,7 +70,6 @@ export {
   prepend,
   range,
   reject,
-  reduceLazy,
   repeat,
   reverse,
   scan,
