+++ website/function.json
@@ -26,7 +26,6 @@
     "pluck",
     "prepend",
     "range",
-    "reduceLazy",
     "reject",
     "repeat",
     "reverse",
@@ -97,6 +96,7 @@
     "prop",
     "props",
     "reduce",
+    "reduceLazy",
     "size",
     "some",
     "sort",
